aider: Added instructions on how to create a standalone executable using PyInstaller in the README.md file.
# Aider chat conversation:

USER: Is there a way to create a separate binary that can run without python on windows

ASSISTANT: Yes, there is a way to create a standalone executable file from a Python script that can run without Python being installed on the system. This can be achieved using a tool called PyInstaller.

However, please note that PyInstaller is not a part of the standard Python library, so it needs to be installed separately. You can install it using pip:

```bash
pip install pyinstaller
```

Once PyInstaller is installed, you can use it to convert your Python script into an executable. Here is the basic command:

```bash
pyinstaller --onefile your_script.py
```

The `--onefile` option tells PyInstaller to create a single executable file. If you don't use this option, PyInstaller will create a directory that contains an executable along with some other files that the executable needs to run.

Please note that the executable will be created in the `dist` directory that PyInstaller creates in the same directory where your Python script is located.

Also, please note that the executable will be platform-specific. That means if you create the executable on a Windows system, it will only run on Windows. If you want to create an executable for a different platform, you need to run PyInstaller on that platform.

Please let me know if you need help modifying your Python script or if you have any other questions.

USER: can you provide that info in the readme under usage

ASSISTANT: Sure, I can provide instructions on how to convert the Python script into a standalone executable using PyInstaller in the README.md file.

Here is the proposed change:

```markdown
README.md
<<<<<<< HEAD
=======
## Usage

This script can be run as a standalone executable without needing Python installed on the system. Here are the steps to create the executable:

1. Install PyInstaller. PyInstaller is not a part of the standard Python library, so it needs to be installed separately. You can install it using pip:

```bash
pip install pyinstaller
```

2. Once PyInstaller is installed, you can use it to convert the Python script into an executable. Here is the basic command:

```bash
pyinstaller --onefile your_script.py
```

The `--onefile` option tells PyInstaller to create a single executable file. If you don't use this option, PyInstaller will create a directory that contains an executable along with some other files that the executable needs to run.

The executable will be created in the `dist` directory that PyInstaller creates in the same directory where your Python script is located.

Please note that the executable will be platform-specific. That means if you create the executable on a Windows system, it will only run on Windows. If you want to create an executable for a different platform, you need to run PyInstaller on that platform.
>>>>>>> updated
```

Please confirm if you want to proceed with this change.